+++ CHANGELOG.md
@@ -1,6 +1,24 @@
 
 # Changelog
 
+## Version 0.9.5
+
+- Further API compatibility and ergonomics improvements.
+
+## Version 0.9.4
+
+- Make `api.new.get.email_id()` able to work with DOMEmail directly.
+
+## Version 0.9.3
+
+- Fix bug introduced in `api.new.get.email_data()` in version 0.9.2
+  when not providing email-id.
+
+## Version 0.9.2
+
+- Make `api.new.get.email_data()` handle legacy-style IDs when
+  present to improve compatibility. Creates console-warning when detected.
+
 ## Version 0.9.1
 
 - Add ability to get new-style email-id directly from DOM element in